# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "5.3.0",
    "title": "Hufflepuff Bob's Computer Repair Shop API for Group 2",
    "description": "BCRS API Documentation for managing users and invoices.\n\n\n\nGroup Github Repo:\n\n*[BCRS repository](https://github.com/BrookeTaylor/bcrs)"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "API for users on the company website"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "findAll",
        "description": "Get a list of users",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {

                    "firstName": {
                      "type": "string",
                      "example": "Johann"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Bach"
                    },
                    "email": {
                      "type": "string",
                      "example": "bach@nodebucket.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "Password01"
                    },
                    "phoneNumber": {
                      "type": "number",
                      "example": "8001234567"
                    },
                    "address": {
                      "type": "string",
                      "example": "1000 Galvin Rd S, Bellevue, NE 68005"
                    },
                    "securityQuestions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": "string",
                          "answer": "string"
                        }
                      }
                    },
                    "role": {
                      "type": "string",
                      "example": "standard"
                    },
                    "isDisabled": {
                      "type": "string",
                      "example": "false"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{empId}": {
      "get": {
        "tags": ["Users"],
        "summary": "findByUserId",
        "description": "Get a user by their valid empId.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "description": "A valid empId.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "empId": {
                      "type": "number",
                      "example": "1008"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Johann"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Bach"
                    },
                    "email": {
                      "type": "string",
                      "example": "bach@nodebucket.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "Password01"
                    },
                    "phoneNumber": {
                      "type": "number",
                      "example": "8001234567"
                    },
                    "address": {
                      "type": "string",
                      "example": "1000 Galvin Rd S, Bellevue, NE 68005"
                    },
                    "securityQuestions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": "string",
                          "answer": "string"
                        }
                      }
                    },
                    "role": {
                      "type": "string",
                      "example": "standard"
                    },
                    "isDisabled": {
                      "type": "boolean",
                      "example": "false"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{empId}": {
      "delete": {
        "tags": ["Users"],
        "summary": "deleteUser",
        "description": "Delete a user by their empId.",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "description": "A valid empId.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }

  }
}